Added an extra sanity check to the wiki mirroring script to avoid uploading to the mirror if a segmentation fault is detected in httrack
git-svn-id: https://rpg.hamsterrepublic.com/source/web@1049 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ mirror/wiki/mirror.sh
@@ -32,7 +32,14 @@ httrack --quiet --extended-parsing=YES \
   +"*/index.php?title=MediaWiki:Monobook.css&*" \
   +"*/index.php?title=-*gen=css*" \
   +"*/favicon.ico" \
-  > /dev/null
+  > httrack.out 2>&1
+
+SEGFAULT=`grep "Segmentation fault" httrack.out`
+rm httrack.out
+if [ -s "${SEGFAULT}" ] ; then
+  echo "SANITY CHECK FAILED: Segmentation fault while downloading"
+  exit 1
+fi
 
 echo "manually download Browser-specific files"
 wget -q -O gilgamesh.hamsterrepublic.com/wiki/ohrrpgce/skins/common/IEFixes.js http://gilgamesh.hamsterrepublic.com/wiki/ohrrpgce/skins/common/IEFixes.js
